[MOD] page.tsx 파일에 있는 코드들 역활 분리 [ADD] 각각 역활에 따른 폴더에  파일 추가
