+++ src/hooks/home_page/useSelectForm.ts
@@ -4,7 +4,6 @@ export default function useSelectForm() {
   const [selectSee, setSelectSee] = useState<string>('시 선택');
   const [selectGunGue, setSelectGunGue] = useState<string>('군 /구 선택');
 
-  // center 좌표에 다음과 같이 주스탠드에서 받아온 데이터를 대입
   const [selectState, setSelectState] = useState({
     center: { lat: 37.566535, lng: 126.9779692 },
     isPanto: false,
