+++ src/util/home_page/getDistance.ts
@@ -1,0 +1,11 @@
+export default function getDistance(lat1: number, lon1: number, lat2: number, lon2: number): number {
+  const R = 6371; // 지구의 반지름 (단위: km)
+  const dLat = (lat2 - lat1) * (Math.PI / 180);
+  const dLon = (lon2 - lon1) * (Math.PI / 180);
+  const a =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(lat1 * (Math.PI / 180)) * Math.cos(lat2 * (Math.PI / 180)) * Math.sin(dLon / 2) * Math.sin(dLon / 2);
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+  const distance = R * c; // 두 지점 사이의 거리 (단위: km)
+  return distance;
+}
